+++ src/Administration/Resources/config/services.xml
@@ -149,6 +149,13 @@
             <argument type="service" id="cache.object" />
         </service>
 
+        <service id="Shopware\Administration\Snippet\AppLifecycleSubscriber">
+            <argument type="service" id="Shopware\Core\Framework\App\Source\SourceResolver"/>
+            <argument type="service" id="Shopware\Administration\Snippet\AppAdministrationSnippetPersister"/>
+
+            <tag name="kernel.event_subscriber"/>
+        </service>
+
         <service id="Shopware\Administration\Framework\Routing\KnownIps\KnownIpsCollector"/>
 
         <!-- @deprecated tag:v6.8.0 Will be removed -->
